Analytics Setup
Added required code to default.html for analytics to work. Also added an analytics.html file to the includes to attach it to every page that is built. This requires the site to be built like "JEKYLL_ENV=production jekyll build" or the logic in the default.html will not work.